Finish previous simplication
Alter the output of "query base" to be, I think, a little clearer,
and to report the Init id explicitly - I believe this is more obvious
that implicitly (and only) reporting it if a merge/push was not
found.

Also, finish "localising" looking up said Init to as high in the
call tree as seems sensible.